+++ CommonTasks/ScopeOfCalculation.cs
@@ -3,8 +3,8 @@
     public class ScopeOfCalculation
     {
 
-        public int Start { get; set; }
-        public int End { get; set; }
+        public long Start { get; set; }
+        public long End { get; set; }
 
 
     }
